+++ src/GSS.Authentication.CAS.AspNetCore/TicketStoreWrapper.cs
@@ -1,17 +1,17 @@
 using System;
-using System.Security.Principal;
+using System.Security.Claims;
 using System.Threading.Tasks;
 using GSS.Authentication.CAS.Security;
-using Microsoft.Owin.Security;
-using Microsoft.Owin.Security.Cookies;
+using Microsoft.AspNetCore.Authentication;
+using Microsoft.AspNetCore.Authentication.Cookies;
 
-namespace GSS.Authentication.CAS.Owin
+namespace GSS.Authentication.CAS.AspNetCore
 {
-    public class AuthenticationSessionStoreWrapper : IAuthenticationSessionStore
+    public class TicketStoreWrapper : ITicketStore
     {
         private readonly IServiceTicketStore _store;
 
-        public AuthenticationSessionStoreWrapper(
+        public TicketStoreWrapper(
             IServiceTicketStore store)
         {
             _store = store;
@@ -42,30 +42,44 @@ namespace GSS.Authentication.CAS.Owin
 
         private static ServiceTicket BuildServiceTicket(AuthenticationTicket ticket)
         {
-            var identity = ticket.Identity;
+            var ticketId = ticket.Properties.GetTokenValue("access_token") ?? Guid.NewGuid().ToString();
+            var principal = ticket.Principal;
             var properties = ticket.Properties;
-            var ticketId = properties?.GetServiceTicket() ?? Guid.NewGuid().ToString();
-            var assertion = (identity as CasIdentity)?.Assertion
+            var assertion = (principal as CasPrincipal)?.Assertion
+                ?? (principal.Identity as CasIdentity)?.Assertion
                 ?? new Assertion(
-                    identity.GetPrincipalName(),
-                    null,
-                    properties?.IssuedUtc,
-                    properties?.ExpiresUtc);
-            return new ServiceTicket(ticketId, assertion, identity.Claims, identity.AuthenticationType);
+                    principal.GetPrincipalName(),
+                    null, properties.IssuedUtc,
+                    properties.ExpiresUtc);
+            return new ServiceTicket(ticketId, assertion, principal.Claims, principal.Identity.AuthenticationType);
         }
 
         private static AuthenticationTicket BuildAuthenticationTicket(ServiceTicket ticket)
         {
             var assertion = ticket.Assertion;
-            var identity = new CasIdentity(assertion, ticket.AuthenticationType);
+            var principal = new CasPrincipal(assertion, ticket.AuthenticationType);
+            if (principal.Identity is not ClaimsIdentity identity)
+            {
+                return new AuthenticationTicket(
+                   principal,
+                   new AuthenticationProperties
+                   {
+                       IssuedUtc = assertion.ValidFrom,
+                       ExpiresUtc = assertion.ValidUntil
+                   },
+                   ticket.AuthenticationType);
+            }
+
             identity.AddClaims(ticket.Claims);
+
             return new AuthenticationTicket(
-                identity,
+                principal,
                 new AuthenticationProperties
                 {
                     IssuedUtc = assertion.ValidFrom,
                     ExpiresUtc = assertion.ValidUntil
-                });
+                },
+                ticket.AuthenticationType);
         }
     }
 }
