+++ src/GSS.Authentication.CAS.Core/GSS.Authentication.CAS.Core.csproj
@@ -3,6 +3,7 @@
   <PropertyGroup>
     <TargetFramework>netstandard2.0</TargetFramework>
     <Description>Core components for CAS Authentication</Description>
+    <RootNamespace>GSS.Authentication.CAS</RootNamespace>
   </PropertyGroup>
 
   <ItemGroup>
