+++ src/GSS.Authentication.CAS.Core/IServiceTicketStore.cs
