+++ src/GSS.Authentication.CAS.Core/ServiceTicket.cs
