+++ src/GSS.Authentication.CAS.Owin/AuthenticationSessionStoreExtensions.cs
