+++ src/GSS.Authentication.CAS.Owin/Events/ICasAuthenticationProvider.cs
@@ -1,5 +1,7 @@
 using System.Threading.Tasks;
+using Microsoft.Owin.Security;
 
+// ReSharper disable once CheckNamespace
 namespace GSS.Authentication.CAS.Owin
 {
     /// <summary>
