+++ test/GSS.Authentication.CAS.AspNetCore.Tests/CasAuthenticationMiddlewareTest.cs
@@ -3,7 +3,6 @@ using System.Linq;
 using System.Net;
 using System.Net.Http;
 using System.Security.Claims;
-using System.Security.Principal;
 using System.Threading;
 using System.Threading.Tasks;
 using GSS.Authentication.CAS.Security;
