+++ test/GSS.Authentication.CAS.AspNetCore.Tests/CasSingleSignOutMiddlewareTest.cs
@@ -6,7 +6,6 @@ using System.Text.Json;
 using System.Threading.Tasks;
 using GSS.Authentication.CAS.Testing;
 using Microsoft.AspNetCore.Authentication.Cookies;
-using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.TestHost;
 using Microsoft.Extensions.FileProviders;
