+++ test/GSS.Authentication.CAS.Core.Tests/Cas20ServiceTicketValidationTest.cs
@@ -2,15 +2,15 @@ using System;
 using System.Net;
 using System.Net.Http;
 using System.Security.Authentication;
-using System.Security.Principal;
 using System.Threading;
 using System.Threading.Tasks;
+using GSS.Authentication.CAS.Testing;
 using GSS.Authentication.CAS.Validation;
 using Microsoft.Extensions.FileProviders;
 using RichardSzalay.MockHttp;
 using Xunit;
 
-namespace GSS.Authentication.CAS.Testing.Validation
+namespace GSS.Authentication.CAS.Core.Tests
 {
     public class Cas20ServiceTicketValidationTest : IClassFixture<CasFixture>
     {
