+++ test/GSS.Authentication.CAS.Owin.Tests/FakeDataProtectionProvider.cs
