+++ test/GSS.Authentication.CAS.Testing/FileProviderExtensions.cs
@@ -4,8 +4,9 @@ using System.Net.Http;
 using System.Net.Http.Headers;
 using System.Text.Json;
 using System.Threading.Tasks;
+using Microsoft.Extensions.FileProviders;
 
-namespace Microsoft.Extensions.FileProviders
+namespace GSS.Authentication.CAS.Testing
 {
     public static class FileProviderExtensions
     {
